+++ sonar-project.properties
@@ -85,6 +85,8 @@ sonar.typescript.tsconfigPath=tsconfig.json
 # - session 페이지: SSE 스트리밍 핸들러가 서비스마다 동일 Next.js 패턴
 # - DailyFortune: 카드 플립 패턴이 CardFace/CardBack 등 다른 컴포넌트와 중복됨
 sonar.cpd.exclusions=\
+  src/hooks/useUserInfoForm.ts,\
+  src/components/common/BirthTimeInput.tsx,\
   src/components/common/PageSpinner.tsx,\
   src/components/effects/ThemeEffectEngine.tsx,\
   src/components/effects/ThemeAtmosphereLayer.tsx,\
